Move UrlText and small updates
